Update patch set 4
Patch Set 4:

> > (1 comment)
 > 
 > I did drill down in a few places to be sure it had arm64 packages. 
 > It does. However, as we found out there does not exist one for
 > jdk9...only 8.  What do you think of fetching specific deb's and
 > installing with dpkg or something along those lines for arm?

Yes that is fine. Java 9 is not officially supported on any distros as far as I can tell. That's why for Ubuntu we install that PPA. So if we need to fetch a dpkg for ARM systems that's fine imo.

Patch-set: 4
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0