Remove OpenJDK 9 from CentOS image builds
This currently breaks everyone trying to use a CentOS 7 builder as
the repo seems to have disappeared since August 23rd, 2018 (last
known time it worked in OpenDaylight).

This breaks everyone so disable it until we have a new solution to
get OpenJDK 9 installed on CentOS 7 systems.

Officially RedHat (Ref) has stated that they are skipping support
of OpenJDK 9 and 10 and will support 11. We should probably follow
their lead on this one.

Ref: https://access.redhat.com/articles/1299013
Issue: RELENG-1175
Change-Id: Ief7710b6469833423156c9695681379b12d349cb
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>